Se trabaja en el controller se crean metodos para inicializar y para seleccionar la casilla hay error ya que no consigo enlazar el evento a los botones

+++ CristianNauzetTCMarzo/src/es/iespuertodelacruz/crisnau/cristiannauzettcmarzo/model/Punto.java
@@ -24,12 +24,14 @@ public class Punto {
         this.y = y;
     }
 
+
+
     //Getters
     public int getX() {
         return x;
     }
 
-    public double getY() {
+    public int getY() {
         return y;
     }
